진짜 진짜 마지막/토큰 관련 오류 해결

+++ src/main/java/com/example/loginlivesession2/domain/Post.java
@@ -23,8 +23,8 @@ public class Post extends Timestamped {
 
     @Column(nullable = false)
     private String nickname;
-
-    private long postlike;
+    
+    private int postlike;
 
     @Column(nullable = false)
     private String title;
