+++ src/main/java/com/example/loginlivesession2/dto/response/PostAllResDto.java
@@ -13,7 +13,7 @@ public class PostAllResDto {
     private LocalDateTime createAt;
     private String nickname;
     private long postid;
-    private Long like;
+    private int like;
     private String title;
     private String content;
